Drop dirport use of endosome
Reading from the DirPort is simple and endosome doesn't make it any simpler.

This 'context' stuff is making my head hurt. Seems there's three: tcp, ssl, and
link. The tcp context is only used by this DirPort demo script so dropping it
so we can focus on the later two.